requirements.txt
python-telegram-bot==21.*
anthropic
playwright
recipe-scrapers
kptncook
python-dotenv
httpx
